put parser dto into dto file
