+++ backend/src/dto/dto.ts
@@ -41,6 +41,7 @@ import {
   UpdateTagSchema,
   VerifyApiKeySchema,
 } from "@repo/common";
+import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";
 import { createZodDto } from "nestjs-zod";
 // This file wraps the Objects into an DTO Swagger can see.
 
@@ -110,3 +111,15 @@ export class UpdateMediaItemDto extends createZodDto(UpdatedMediaItemSchema) {}
 export class MediaCropDto extends createZodDto(MediaCropSchema) {}
 export class CreateMediaCropDto extends createZodDto(CreateMediaCropSchema) {}
 export class UpdateMediaCropDto extends createZodDto(UpdatedMediaCropSchema) {}
+
+// CSV Upload DTO
+export class ParserUploadCsvBodyDto {
+  @ApiProperty({ type: "string", format: "binary" })
+  file?: unknown;
+
+  @ApiPropertyOptional({
+    description:
+      "Optional fallback path on the server when no multipart file is uploaded.",
+  })
+  filePath?: string;
+}
